PR-793: Fix: language not set when creating new file
Props @Hero2006

Ref: issues/752
Thanks @xylographe

When NPP is invoked with an non-existing path argument, it asks if the file should be created. If confirmed, a new file will be opened, but the language (syntax) will not be set automatically.
Reproduce:
1. Execute command `rm -f foobar.cs; notepad++ foobar.cs` in your favorite shell.
2. Confirm when asked to create new file.
3. Against expectation, the language is not set to C#.

Code:
buf->setLangType(ndds._lang);
doesn't need because buffer has already logic for detecting LangType.
P.S. maybe other strings are not needed too...